Funcionando Todo lo SQL

+++ app/src/main/java/tepic/tecnm/mx/eamtlmhm/ladm_u3_practica2_firebasefirestore/MainActivity2.kt
@@ -6,6 +6,7 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import kotlinx.android.synthetic.main.activity_main2.*
 import java.text.SimpleDateFormat
 import java.util.*
@@ -40,7 +41,7 @@ class MainActivity2 : AppCompatActivity() {
             var intent = getIntent()
             val titulo = intent.getStringExtra("titulo").toString()//Crashea si no existe un Extra
             val texto = intent.getStringExtra("texto").toString()
-            id = intent.getStringExtra("id").toString().toInt()
+            id = intent.getIntExtra("idNOTAS",0)
 
             modificacion = true
             //Coloca el contenido mandado a los txt
@@ -50,22 +51,36 @@ class MainActivity2 : AppCompatActivity() {
     }
 
     private fun actualizar(){
-        val nota = Nota(this)
-        nota.id = id
-        nota.titulo =  txtTitle.text.toString()
-        nota.contenido = txtContenido.text.toString()
-        nota.hora = txtTime.text.toString()
-        nota.fecha =  txtDate.text.toString()
+        AlertDialog.Builder(this)
+            .setMessage("¿SEGURO QUE QUIERE ACTUALIZAR LA NOTA CON ID = ${id}?")
+            .setPositiveButton("ACTUALIZAR"){d,i->
+                val nota = Nota(this)
+                nota.id = id
+                nota.titulo =  txtTitle.text.toString()
+                nota.contenido = txtContenido.text.toString()
+                nota.hora = txtTime.text.toString()
+                nota.fecha =  txtDate.text.toString()
 
-        if(nota.updateNota()){ Toast.makeText(this,"Se Actualizo Exitosamente",Toast.LENGTH_LONG).show()}
-        else{ Toast.makeText(this,"Error al actualizar",Toast.LENGTH_LONG).show()}
+                if(nota.updateNota()){ Toast.makeText(this,"Se Actualizo Exitosamente",Toast.LENGTH_LONG).show()}
+                else{ Toast.makeText(this,"Error al actualizar",Toast.LENGTH_LONG).show()}
+            }
+            .setNegativeButton("CANCELAR"){d,i-> }
+            .show()
+
     }//end actualizar
 
     private fun borrarNotas() {
-        val nota = Nota(this)
-        nota.id = id
-        if(nota.deleteNota()){ Toast.makeText(this,"Se elimino exitosamente",Toast.LENGTH_LONG).show()}
-        else{ Toast.makeText(this,"Error al eliminar",Toast.LENGTH_LONG).show()}
+        AlertDialog.Builder(this)
+            .setMessage("¿SEGURO QUE QUIERE ELIMINAR LA NOTA CON ID = ${id}?")
+            .setPositiveButton("ELIMINAR"){d,i->
+                val nota = Nota(this)
+                nota.id = id
+                if(nota.deleteNota()){ Toast.makeText(this,"Se elimino exitosamente",Toast.LENGTH_LONG).show()}
+                else{ Toast.makeText(this,"Error al eliminar",Toast.LENGTH_LONG).show()}
+            }
+            .setNegativeButton("CANCELAR"){d,i-> }
+            .show()
+
     }
 
     private fun guardar(){
@@ -73,15 +88,19 @@ class MainActivity2 : AppCompatActivity() {
             actualizar()
         }//end if
         else {
-            val nota = Nota(this)
-            nota.titulo =  txtTitle.text.toString()
-            nota.contenido = txtContenido.text.toString()
-            nota.hora = txtTime.text.toString()
-            nota.fecha =  txtDate.text.toString()
-            if(nota.insertNota()){
-                Toast.makeText(this,"Nota Agregada Exitosamente",Toast.LENGTH_LONG).show()
+            if(!txtTitle.text.toString().isEmpty() || !txtContenido.text.toString().isEmpty()){
+                val nota = Nota(this)
+                nota.titulo =  txtTitle.text.toString()
+                nota.contenido = txtContenido.text.toString()
+                nota.hora = txtTime.text.toString()
+                nota.fecha =  txtDate.text.toString()
+                if(nota.insertNota()){
+                    Toast.makeText(this,"Nota Agregada Exitosamente",Toast.LENGTH_LONG).show()
+                }else{
+                    Toast.makeText(this,"Error al agregar Nota",Toast.LENGTH_LONG).show()
+                }
             }else{
-                Toast.makeText(this,"Error al agregar Nota",Toast.LENGTH_LONG).show()
+                Toast.makeText(this,"Error rellene titulo y descripcion",Toast.LENGTH_LONG).show()
             }
         }//end else
     }//end guardar
